Improve usage of routes and Link
