+++ src/views/LogisticsEmergency/styles.module.css
@@ -12,7 +12,7 @@
             grid-template-columns: repeat(auto-fit, minmax(13rem, 1fr));
 
             .image {
-                min-height: 10rem;
+                height: 16rem;
             }
         }
         .inner-content {
